+++ Amap.h
@@ -54,13 +54,13 @@
 #define MIN3(a,b,c) (MIN(a,MIN(b,c)))
 #endif
 
-extern double Kmeans(double *src, unsigned char *label, unsigned char *mask, int NI, int n_clusters, double *voxelsize, int *dims, int thresh_mask, int thresh_kmeans, int iters_nu, int pve, double bias_fwhm);
-extern void Amap(double *src, unsigned char *label, unsigned char *prob, double *mean, int nc, int niters, int sub, int *dims, int pve, double weight_MRF, double *voxelsize, int niters_ICM, double offset);
-extern void Pve5(double *src, unsigned char *prob, unsigned char *label, double *mean, int *dims);
-extern void Pve6(double *src, unsigned char *prob, unsigned char *label, double *mean, int *dims);
+extern double Kmeans(float *src, unsigned char *label, unsigned char *mask, int NI, int n_clusters, double *voxelsize, int *dims, int thresh_mask, int thresh_kmeans, int iters_nu, int pve, double bias_fwhm);
+extern void Amap(float *src, unsigned char *label, unsigned char *prob, double *mean, int nc, int niters, int sub, int *dims, int pve, double weight_MRF, double *voxelsize, int niters_ICM, double offset);
+extern void Pve5(float *src, unsigned char *prob, unsigned char *label, double *mean, int *dims);
+extern void Pve6(float *src, unsigned char *prob, unsigned char *label, double *mean, int *dims);
 extern void MrfPrior(unsigned char *label, int nc, double *alpha, double *beta, int init, int *dims);
 #ifdef SPLINESMOOTH
-  extern int splineSmooth( double *src, double lambda, double distance, int subsample, double *separations, int *dims);
+  extern int splineSmooth( float *src, double lambda, double distance, int subsample, double *separations, int *dims);
 #endif
 
 struct point {
