+++ Pve.c
@@ -17,7 +17,7 @@
 #include <math.h>
 #include "Amap.h"
 
-void Pve5(double *src, unsigned char *prob, unsigned char *label, double *mean, int *dims)
+void Pve5(float *src, unsigned char *prob, unsigned char *label, double *mean, int *dims)
 {
   int x,y,z,i,z_area,y_dims,ind,mxi;
   double w, mx;
@@ -58,7 +58,7 @@ void Pve5(double *src, unsigned char *prob, unsigned char *label, double *mean, 
           label[ind] = 255;
           break;
         case GMCSFLABEL: /* GMCSF */
-          w = (src[ind] - mean[CSFLABEL-1])/(mean[GMLABEL-1]-mean[CSFLABEL-1]);
+          w = ((double)src[ind] - mean[CSFLABEL-1])/(mean[GMLABEL-1]-mean[CSFLABEL-1]);
           if(w > 1.0) w = 1.0; if(w < 0.0) w = 0.0;
           new_val[CSFLABEL-1] = (unsigned char) ROUND(255.0*(1-w));
           new_val[GMLABEL-1]  = (unsigned char) ROUND(255.0*w);
@@ -66,7 +66,7 @@ void Pve5(double *src, unsigned char *prob, unsigned char *label, double *mean, 
           label[ind] = ROUND(255.0/3.0*(1.0 + w));
           break;
         case WMGMLABEL: /* WMGM */
-          w = (src[ind] - mean[GMLABEL-1])/(mean[WMLABEL-1]-mean[GMLABEL-1]);
+          w = ((double)src[ind] - mean[GMLABEL-1])/(mean[WMLABEL-1]-mean[GMLABEL-1]);
           if(w > 1.0) w = 1.0; if(w < 0.0) w = 0.0;
           new_val[CSFLABEL-1] = 0;
           new_val[GMLABEL-1]  = (unsigned char) ROUND(255.0*(1-w));
@@ -88,7 +88,7 @@ void Pve5(double *src, unsigned char *prob, unsigned char *label, double *mean, 
   }  
 }
 
-void Pve6(double *src, unsigned char *prob, unsigned char *label, double *mean, int *dims)
+void Pve6(float *src, unsigned char *prob, unsigned char *label, double *mean, int *dims)
 {
   int x,y,z,i,z_area,y_dims,ind,mxi;
   double w, mx;
@@ -129,7 +129,7 @@ void Pve6(double *src, unsigned char *prob, unsigned char *label, double *mean, 
           label[ind] = 255;
           break;
         case BKGCSFLABEL+1: /* BKGCSF */
-          w = src[ind]/mean[CSFLABEL];
+          w = (double)src[ind]/mean[CSFLABEL];
           if(w > 1.0) w = 1.0; if(w < 0.0) w = 0.0;
           new_val[CSFLABEL] = (unsigned char) ROUND(255.0*w);
           new_val[GMLABEL]  = 0;
@@ -137,7 +137,7 @@ void Pve6(double *src, unsigned char *prob, unsigned char *label, double *mean, 
           label[ind] = ROUND(255.0/3.0*w);
           break;
         case GMCSFLABEL+1: /* GMCSF */
-          w = (src[ind] - mean[CSFLABEL])/(mean[GMLABEL]-mean[CSFLABEL]);
+          w = ((double)src[ind] - mean[CSFLABEL])/(mean[GMLABEL]-mean[CSFLABEL]);
           if(w > 1.0) w = 1.0; if(w < 0.0) w = 0.0;
           new_val[CSFLABEL] = (unsigned char) ROUND(255.0*(1-w));
           new_val[GMLABEL]  = (unsigned char) ROUND(255.0*w);
@@ -145,7 +145,7 @@ void Pve6(double *src, unsigned char *prob, unsigned char *label, double *mean, 
           label[ind] = ROUND(255.0/3.0*(1.0 + w));
           break;
         case WMGMLABEL+1: /* WMGM */
-          w = (src[ind] - mean[GMLABEL])/(mean[WMLABEL]-mean[GMLABEL]);
+          w = ((double)src[ind] - mean[GMLABEL])/(mean[WMLABEL]-mean[GMLABEL]);
           if(w > 1.0) w = 1.0; if(w < 0.0) w = 0.0;
           new_val[CSFLABEL] = 0;
           new_val[GMLABEL]  = (unsigned char) ROUND(255.0*(1-w));
